requirements.txt
openai==1.39.0
python-dotenv==1.0.1
streamlit==1.37.0
azure-cosmos==4.7.0
plotly==5.23.0
pandas==2.2.2
langchain==0.2.12
langchain-openai==0.1.20
langchain-community==0.2.11
azure-search-documents==11.5.1
azure-identity==1.17.1
